# PERF: groupby(...).__len__