- [ ] closes #xxxx (Replace xxxx with the GitHub issue number)
- [ ] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [ ] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [ ] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [ ] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

Enabled by #55738. Prior to that, `ngroups` was not reliable.

```
size = 100_000
df = pd.DataFrame(
    {
        "a": np.random.randint(0, 100, size),
        "b": np.random.randint(0, 100, size),
    }
)
%timeit len(df.groupby(["a", "b"]))

# 161 ms ± 962 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)    <-- main
# 4.96 ms ± 229 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  <-- PR
```